+++ app/src/main/java/br/com/transporte/AppGhn/ui/fragment/formularios/FormulariosInterface.java
@@ -26,6 +26,6 @@ public interface FormulariosInterface {
 
     void adicionaObjetoNoBancoDeDados();
 
-    int configuraObjetoNaCriacao();
+    Long configuraObjetoNaCriacao();
 
 }
